# Further prune down JSON recursion for performance.