A small improvement for #301, like #612.

| test              | 2d612e1 | ec3e8b6 |
|-------------------|---------|---------|
| update_list       | 1.55s   | 1.62s   |
| update_list_gnome | 6.27s   | 2.26s   |
| update_view       | 0.68s   | 0.67s   |
| update_view_gnome | 4.32s   | 1.27s   |
| release_list      | 0.02s   | 0.02s   |
| release_view      | 0.02s   | 0.01s   |
| comment_list      | 4.47s   | 1.96s   |
| comment_view      | 0.05s   | 0.04s   |
| user_list         | 0.18s   | 0.33s   |
| user_view         | 0.11s   | 0.08s   |